+++ naturtag/metadata/image_metadata.py
