--- naturtag/models/__init__.py
@@ -1,8 +1,0 @@
-""" Data classes with utilities for managing specific resources """
-# flake8: noqa: F401
-
-from pyinaturalist.models import Observation
-
-from naturtag.models.image_metadata import ImageMetadata
-from naturtag.models.keyword_metadata import HIER_KEYWORD_TAGS, KEYWORD_TAGS, KeywordMetadata
-from naturtag.models.meta_metadata import MetaMetadata
